requirements.txt
telethon==1.37.0
tortoise-orm==0.20.0
aerich[sqlite]==0.7.2
asyncpg==0.29.0
python-decouple==3.8
uvloop==0.19.0
apscheduler==3.10.4
redis[hiredis]==5.0.1